pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slenps"
version = "0.1.0"
description = "Some Pipelines"
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
requires-python = ">=3.8"

dependencies = [
    "scipy==1.10.1",
    "numpy",
    "pandas",
    "gensim",
    "psutil",
    "scikit-learn",
    "sentence-transformers",
    "tqdm",
    "umap-learn"
]

[tool.setuptools_scm]
write_to = "slenps/version.py"